Update patch set 1
Patch Set 1:

The Samsung implementation of bcmdhd driver in the kernel is very different from reference broadcom implementations especially on smdk4412. This makes it hard to properly configure the driver as builtin because upon reset, the wifi driver fails to load again and also there are a lot of MMC related errors which occur. Furthermore the MAC and NVRAM is also screwed up since NVRAM for the same device might be different based on the WiFi chip that is used. 

There are a lot of devices that use the BOARD_HAVE_SAMSUNG_WIFI flag and it's a lot of work to get the WiFi working as builtin on all the devices.

Patch-set: 1