Update patch set 1
Patch Set 1:

I'm not sure I like the idea of autoloading the module in the background. Might be better to check if the wlan interface actually exists, and stop further processing if it doesn't. I'm also not sure if the wifi hal is actually of any use or benefit for old chips like the bcm4329. Heck, I don't even know what it is for actually. :)

Patch-set: 1