Update patch set 1
Patch Set 1: Verified+1

we had a service p2p in MM and N
https://github.com/Suicide-Squirrel/android_device_lge_g4-common-LOS/blob/d61f2184deb753700dc033be00d6244a7a70c349/rootdir/etc/init.qcom.rc#L888

in O this was removed and p2p was added in wpa
https://github.com/Suicide-Squirrel/android_device_lge_g4-common-LOS/blob/0b20553019d3f81e876dde53f6b78decd317b102/rootdir/etc/init.qcom.rc#L999

In O we used to get an error that :
p2p0 device is missing.

So this commit fixes that and we have wifi direct again. (for devices using brcm wifi)

*Sorry for using this repo to point to the lines, but i haven't updated g4 officially since i haven't enforced it

Patch-set: 1
Reviewer: Pavlos Theodorou <18258@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=+1